reading from cli and file
