+++ ems/driver.py
@@ -1,5 +1,5 @@
-from ems.config_reader.commandline import SimulatorCLI
-from ems.config_reader.reader import ResolveConfigs
+from ems.config_reader.loaders import UserArguments
+from ems.config_reader.resolver import ResolveConfigs
 
 # TODO: Goal of this file is to abstract out a lot of the run.py in the above directory.
 # TODO: user inputs from CLI args (parse_args) and configurations stored in the files.
@@ -10,9 +10,10 @@ class UserInput: # TODO Request for comments (RFC) on name
 
     def __init__(self):
         """
-
+        This class reads from two sources of truth and resolves them
+        Command line arguments always overrides file configurations
         """
-        command_line_args = SimulatorCLI().args()
+        command_line_args = UserArguments()
         settings = ResolveConfigs(debug=command_line_args.debug, args=command_line_args)
         self.args = settings
